Prepare the stable/wallaby jobs with pinned manila-tempest-plugin
As stable/wallaby is in EM phase, and currently broken
with the Tempest master. We are pinning tempest there (depends-on).

We are using the tempest 29.0.0 on stable/wallaby (depends-on) and
compatible tag for tempest 29.0.0 is manila-tempest-plugin 1.8.0
(for ddt version compatible with stable/wallaby constraints and tempest
version)
- https://github.com/openstack/manila-tempest-plugin/blob/1.8.0/requirements.txt
- https://github.com/openstack/requirements/blob/stable/wallaby/upper-constraints.txt#L80

Otherwise it fail for ddt incompatible version
- https://zuul.opendev.org/t/openstack/build/4dad3a22fe5d485a96722cb6d68da63d

Preparing the new set of jobs with manila-tempest-plugin pin to 1.8.0
which will be used to test the stable/wallaby.

Needed-By: https://review.opendev.org/c/openstack/manila/+/871864

Depends-On: https://review.opendev.org/c/openstack/devstack/+/871782
Change-Id: Ibf8c885529098417e8ff9b417bf172218f8d684a